수정: 멘토는 팀 탈퇴 버튼이 비활성화 되도록 변경 (#187)

+++ src/main/java/io/seoul/helper/service/TeamService.java
@@ -170,6 +170,8 @@ public class TeamService {
         Team team = findTeam(id);
         Member member = memberRepo.findMemberByTeamAndUser(team, user)
                 .orElseThrow(() -> new Exception("Not this team member"));
+        if (team.getStatus() != TeamStatus.READY)
+            throw new Exception("This team is already running");
         team.outTeam();
         teamRepo.save(team);
         memberRepo.delete(member);
